Added support for Vikendi
Sorry for another pull request, wanted to add the new map. Its 176 tiles, the scale is 12, the indices aren't the same as other maps so I created another variable "indicespos" to account for this. The path is UmodelExport\Maps\DihorOtok\Art\Heightmap. Tested it, every map works, nothing is broken.